requirements.txt
fastapi
sqlalchemy
alembic
python-dotenv
uvicorn
psycopg2-binary
passlib
bcrypt
python-jose